+++ util/retry/retry.go
@@ -1,15 +1,9 @@
 package retry
 
 import (
-	"net"
-	"net/url"
-	"strings"
 	"time"
 
-	apierr "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/util/wait"
-
-	argoerrs "github.com/argoproj/argo/errors"
 )
 
 // DefaultRetry is a default retry backoff settings when retrying API calls
@@ -18,48 +12,4 @@ var DefaultRetry = wait.Backoff{
 	Duration: 10 * time.Millisecond,
 	Factor:   1.0,
 	Jitter:   0.1,
-}
-
-func IsTransientErr(err error) bool {
-	if err == nil {
-		return false
-	}
-	err = argoerrs.Cause(err)
-	return isExceededQuotaErr(err) || apierr.IsTooManyRequests(err) || isResourceQuotaConflictErr(err) || isTransientNetworkErr(err)
-}
-
-func isExceededQuotaErr(err error) bool {
-	return apierr.IsForbidden(err) && strings.Contains(err.Error(), "exceeded quota")
-}
-
-func isResourceQuotaConflictErr(err error) bool {
-	return apierr.IsConflict(err) && strings.Contains(err.Error(), "Operation cannot be fulfilled on resourcequota")
-}
-
-func isTransientNetworkErr(err error) bool {
-	switch err.(type) {
-	case net.Error:
-		switch err.(type) {
-		case *net.DNSError, *net.OpError, net.UnknownNetworkError:
-			return true
-		case *url.Error:
-			// For a URL error, where it replies back "connection closed"
-			// retry again.
-			if strings.Contains(err.Error(), "Connection closed by foreign host") {
-				return true
-			}
-		default:
-			if strings.Contains(err.Error(), "net/http: TLS handshake timeout") {
-				// If error is - tlsHandshakeTimeoutError, retry.
-				return true
-			} else if strings.Contains(err.Error(), "i/o timeout") {
-				// If error is - tcp timeoutError, retry.
-				return true
-			} else if strings.Contains(err.Error(), "connection timed out") {
-				// If err is a net.Dial timeout, retry.
-				return true
-			}
-		}
-	}
-	return false
 }
